refactor: improve stepper types and simplify step actions props
