+++ src/ui/component/stepper/stepper.tsx
@@ -2,23 +2,14 @@ import { TransitionGroup, CSSTransition } from 'react-transition-group'
 import { type FC } from 'react'
 import type { Optional } from '@/util/type'
 import { StepperProgress } from './stepperProgress/stepperProgress'
-import type { StepperControls } from './useStepper'
+import type { Step, UseStepper } from './useStepper'
 import { useStepper } from './useStepper'
 import { StepperActions } from './stepperActions/stepperActions'
 import './stepper.scss'
 
-export type StepId = string
-
-export type StepStatus = 'complete' | 'incomplete'
-
-export type Step = {
-  id: StepId
-  order: number
-  status: StepStatus
+export type StepElement = Optional<Omit<Step, 'order'>, 'status'> & {
   content: JSX.Element
 }
-
-export type StepElement = Optional<Omit<Step, 'order'>, 'status'>
 
 type StepperProps = {
   steps: StepElement[]
@@ -26,7 +17,7 @@ type StepperProps = {
 
 // eslint-disable-next-line max-lines-per-function
 export const Stepper: FC<StepperProps> = ({ steps: stepElements }) => {
-  const { steps, nextStep, previousStep, activeStepId, previousActiveStepId }: StepperControls =
+  const { steps, nextStep, previousStep, activeStepId, previousActiveStepId }: UseStepper =
     useStepper(
       stepElements.map(({ id, status }) => ({
         id,
@@ -34,7 +25,8 @@ export const Stepper: FC<StepperProps> = ({ steps: stepElements }) => {
       }))
     )
 
-  const activeStep = stepElements.find(step => step.id === activeStepId) ?? stepElements[0]
+  const activeStepElement = stepElements.find(step => step.id === activeStepId) ?? stepElements[0]
+  const activeStep = steps.find(step => step.id === activeStepId) ?? steps[0]
 
   return (
     <div className="okp4-dataverse-portal-stepper-main">
@@ -47,15 +39,15 @@ export const Stepper: FC<StepperProps> = ({ steps: stepElements }) => {
       <TransitionGroup className="okp4-dataverse-portal-stepper-content-wrapper">
         {/* timeout transition duration must match the $transition-duration variable from stepper.scss and progressBar.scss */}
         <CSSTransition classNames="transition" key={activeStepId} timeout={600}>
-          <div className="okp4-dataverse-portal-stepper-content">{activeStep.content}</div>
+          <div className="okp4-dataverse-portal-stepper-content">{activeStepElement.content}</div>
         </CSSTransition>
       </TransitionGroup>
 
       <StepperActions
-        activeStepId={activeStepId}
+        isFirstStep={activeStep.order === 0}
+        isLastStep={activeStep.order === steps.length - 1}
         nextStep={nextStep}
         previousStep={previousStep}
-        steps={steps}
       />
     </div>
   )
